Add security to login required views

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, get_object_or_404, redirect, reverse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 from .models import Product, Category
@@ -75,8 +76,13 @@ def product_details(request, product_id):
     return render(request, 'products/product_details.html', context)
 
 
+@login_required
 def add_product(request):
     """ Adding products to the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have the authority to access this page!')
+        return redirect(reverse('index'))
+
     if request.method == 'POST':
         form = ManageProductsForm(request.POST, request.FILES)
         if form.is_valid():
@@ -96,8 +102,13 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """ Editing Products on the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have the authority to access this page!')
+        return redirect(reverse('index'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ManageProductsForm(request.POST, request.FILES, instance=product)
@@ -120,8 +131,13 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """Delete products from the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'You do not have the authority to access this page!')
+        return redirect(reverse('index'))
+        
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, f'You have deleted {product.name}')
